renamed files
gonna separate api and auth into folders later
renamed files
